requirements.txt
Django
django-cors-headers
django-filter
django-restframework
djangorestframework